Membership of Functor, Foldable, and Traversable.
Since we already have Data.Foldable.toList, toList is no longer
needed.

Note that AVLTree isn't actually a Functor in the categorical sense,
since we require the morphisms we fmap to be order-preserving if we
are to keep the usual BST semantics. Nevertheless, such an instance
is useful, so it is added.